sound refactor
