+++ docs/ure/ui/USpeaker.html
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ure.ui.USpeaker</li>
+<li>ure.ui.sounds.USpeaker</li>
 </ul>
 </li>
 </ul>
