+++ src/main/java/ure/sys/dagger/AppModule.java
@@ -27,10 +27,9 @@ import ure.things.UThingCzar;
 import ure.ui.Icons.Icon;
 import ure.ui.Icons.IconDeserializer;
 import ure.ui.Icons.UIconCzar;
-import ure.ui.USpeaker;
+import ure.ui.sounds.USpeaker;
 
 import javax.inject.Singleton;
-import java.util.Random;
 
 /**
  *
